refactor: Break big `livechatTyped` into smaller files (#35324)
