+++ apps/meteor/app/lib/server/functions/closeOmnichannelConversations.ts
@@ -3,7 +3,7 @@ import { LivechatRooms } from '@rocket.chat/models';
 
 import { callbacks } from '../../../../lib/callbacks';
 import { i18n } from '../../../../server/lib/i18n';
-import { Livechat } from '../../../livechat/server/lib/LivechatTyped';
+import { closeRoom } from '../../../livechat/server/lib/closeRoom';
 import { settings } from '../../../settings/server';
 
 type SubscribedRooms = {
@@ -13,7 +13,7 @@ type SubscribedRooms = {
 
 export const closeOmnichannelConversations = async (user: IUser, subscribedRooms: SubscribedRooms[]): Promise<void> => {
 	const extraQuery = await callbacks.run('livechat.applyRoomRestrictions', {});
-	const roomsInfo = await LivechatRooms.findByIds(
+	const roomsInfo = LivechatRooms.findByIds(
 		subscribedRooms.map(({ rid }) => rid),
 		{},
 		extraQuery,
@@ -22,8 +22,8 @@ export const closeOmnichannelConversations = async (user: IUser, subscribedRooms
 	const comment = i18n.t('Agent_deactivated', { lng: language });
 
 	const promises: Promise<void>[] = [];
-	await roomsInfo.forEach((room: any) => {
-		promises.push(Livechat.closeRoom({ user, room, comment }));
+	await roomsInfo.forEach((room) => {
+		promises.push(closeRoom({ user, room, comment }));
 	});
 
 	await Promise.all(promises);
