+++ apps/meteor/ee/app/livechat-enterprise/server/lib/AutoCloseOnHoldScheduler.ts
@@ -7,7 +7,7 @@ import { MongoInternals } from 'meteor/mongo';
 import moment from 'moment';
 
 import { schedulerLogger } from './logger';
-import { Livechat } from '../../../../../app/livechat/server/lib/LivechatTyped';
+import { closeRoom } from '../../../../../app/livechat/server/lib/closeRoom';
 
 const SCHEDULER_NAME = 'omnichannel_auto_close_on_hold_scheduler';
 
@@ -82,7 +82,7 @@ export class AutoCloseOnHoldSchedulerClass {
 			comment,
 		};
 
-		await Livechat.closeRoom(payload);
+		await closeRoom(payload);
 	}
 
 	private async getSchedulerUser(): Promise<IUser> {
